win32.c: supprss warning
* win32/win32.c (rb_w32_system_tmpdir): supprss a sign-compare
  warning.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@42394 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ win32/win32.c
@@ -490,9 +490,9 @@ rb_w32_system_tmpdir(WCHAR *path, UINT len)
     }
     p = translate_wchar(path, L'\\', L'/');
     if (*(p - 1) != L'/') *p++ = L'/';
-    if (p - path + numberof(temp) >= len) return 0;
+    if ((UINT)(p - path + numberof(temp)) >= len) return 0;
     memcpy(p, temp, sizeof(temp));
-    return p - path + numberof(temp) - 1;
+    return (UINT)(p - path + numberof(temp) - 1);
 }
 
 /* License: Ruby's */
